+++ CHANGELOG.md
@@ -30,3 +30,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+ - AddNode can take an array of nodes